Adapt to ThingTalk 2
Use the new interfaces to parse and serialize ThingTalk, while
paying attention to the "thingtalk_version" query argument in the
Thingpedia API so we serve the right format.

This commit ensures compatibility with old clients if they
specify an old ThingTalk version (which they should all do),
and also transparently migrates existing ThingTalk 1.0 code
to 2.0 internally.